refactor: optimize imports and move extension
